requirements.txt
Flask
flask-cors
nltk
PyMuPDF
google-ai-generativelanguage>=0.6.16,<0.7.0
python-dotenv
httpcore<0.16